Option to define path spec for Redshift lineage generation Fixing s3 urn generation for Redshift copy in lineage generation
