+++ metadata-ingestion/src/datahub/configuration/source_common.py
@@ -1,4 +1,4 @@
-from typing import Dict, Optional
+from typing import Dict, List, Optional
 
 from pydantic import validator
 from pydantic.fields import Field
